練習

+++ CH10_02_SPDY_OkHttp_1.1/src/main/java/com/main/MainActivity.java
@@ -42,7 +42,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void handleMessage(Message msg) {
                 double sum = (Double) msg.obj;
-                setTitle(sum + " %");
+                setTitle((int) sum + " %");
             }
         };
         private Bitmap mBitmap;
